Bug fix: scalb takes int rather than long as parameter

+++ test/jdk/jdk/internal/util/random/RandomSupportTest.java
@@ -27,7 +27,7 @@ public class RandomSupportTest {
     }
     @Test
     public void testNextExponentialSoftCapped() {
-        for (long maxExp = 0; maxExp < 33; maxExp++) {
+        for (int maxExp = 0; maxExp < 33; maxExp++) {
             double max = Math.scalb(1.0, maxExp);
             WorstCaseRandomGenerator rng = new WorstCaseRandomGenerator();
             assertTrue(RandomSupport.computeNextExponentialSoftCapped(rng, max) >= max);
